+++ src/routes/+page.svelte
@@ -19,7 +19,7 @@
   // --------------------------------------------------------------------------
   // DEFAULT PAGE CONTENT - AJDUST TO YOUR NEEDS
   // --------------------------------------------------------------------------
-  const EMAIL = 'mohammad@bnei.dev'; 
+  const EMAIL = 'mohammad@bnei.dev';
 
   const FAQS_PLACEHOLDER = `
 		<h2>What kind of projects do you work on?</h2>
@@ -66,7 +66,8 @@
         data.page?.introStep1 || {
           label: 'BUILD',
           title: 'Crafting Code',
-          description: 'Exploring the process of developing efficient and reliable software solutions.'
+          description:
+            'Exploring the process of developing efficient and reliable software solutions.'
         }
       )
     );
@@ -84,7 +85,8 @@
         data.page?.introStep3 || {
           label: 'LEARN',
           title: 'Sharing Discoveries',
-          description: 'Documenting insights and lessons learned from various technical explorations.'
+          description:
+            'Documenting insights and lessons learned from various technical explorations.'
         }
       )
     );
@@ -173,10 +175,10 @@
 </script>
 
 <svelte:head>
-  <title>Mohammad-Amine BANAEI - My Projects & Knowledge Blog</title>
-  <meta name="description" content="Mohammad-Amine BANAEI's personal blog showcasing projects and technical knowledge." />
-  <link rel="alternate" hreflang="en" href="https://editable.website" />
-  <link rel="canonical" href="https://editable.website" />
+  <title>{title}</title>
+  <meta name="description" content={bio} />
+  <link rel="alternate" hreflang="en" href="https://blog.bnei.dev" />
+  <link rel="canonical" href="https://blog.bnei.dev" />
 </svelte:head>
 
 <WebsiteHeader bind:showUserMenu on:cancel={initOrReset} on:save={savePage}>
@@ -214,8 +216,7 @@
   <div class="max-w-(--breakpoint-md) mx-auto px-6">
     <div class="relative">
       <div class="w-1 bg-gray-900 absolute inset-0 -top-8 bottom-12 mx-auto z-0">
-        <div class="w-4 h-4 rounded-full bg-gray-900 absolute -top-1 -left-[6px]" >
-        </div>
+        <div class="w-4 h-4 rounded-full bg-gray-900 absolute -top-1 -left-[6px]"></div>
       </div>
       <div class="z-10">
         <IntroStep bind:intro={introStep1} />
@@ -228,8 +229,7 @@
       <div class="w-1 bg-gray-900 absolute inset-0 -top-16 bottom-12 mx-auto z-0">
         <div
           class="absolute -bottom-2 -left-[7px] h-0 w-0 border-x-[9px] border-x-transparent border-t-10 border-gray-900"
-        >
-        </div>
+        ></div>
       </div>
     </div>
     <div class="text-center mb-32">
@@ -292,7 +292,6 @@
           alt="Mohammad-Amine BANAEI"
         />
       </div>
-
     </div>
     <div class="">
       <h1 class="text-3xl md:text-5xl font-bold">