+++ vite.config.js
@@ -1,7 +1,8 @@
 import { sveltekit } from '@sveltejs/kit/vite';
+import tailwindcss from '@tailwindcss/vite';
 
 const config = {
-  plugins: [sveltekit()]
+  plugins: [tailwindcss(), sveltekit()]
 };
 
 export default config;