+++ CHANGELOG.md
@@ -6,7 +6,6 @@
 
 ## 0.0.2 (2025-07-23)
 
-
 ### Bug Fixes
 
-* Update email constant to mohammad@bnei.dev ([e895f86](https://github.com/MohammadBnei/editable-blog/commit/e895f869fb1baa758c6d602448e500ced2f92e5d))
+- Update email constant to mohammad@bnei.dev ([e895f86](https://github.com/MohammadBnei/editable-blog/commit/e895f869fb1baa758c6d602448e500ced2f92e5d))